+++ tests/test_visitors/test_ast/test_complexity/test_function/test_arguments_lambda.py
@@ -2,6 +2,7 @@
 
 import pytest
 
+from wemake_python_styleguide.compat.constants import PY38
 from wemake_python_styleguide.visitors.ast.complexity.function import (
     FunctionComplexityVisitor,
     TooManyArgumentsViolation,
@@ -11,6 +12,7 @@ lambda_without_arguments = 'lambda: ...'
 lambda_with_single_argument = 'lambda arg1: ...'
 lambda_with_default_argument = 'lambda arg1=None: ...'
 lambda_with_single_args = 'lambda *args: ...'
+lambda_with_posonly_args = 'lambda arg, /: ...'
 lambda_with_single_kwargs = 'lambda **kwargs: ...'
 lambda_with_single_kwonly = 'lambda *, kwonly=True: ...'
 
@@ -20,6 +22,10 @@ lambda_with_single_kwonly = 'lambda *, kwonly=True: ...'
     lambda_with_single_argument,
     lambda_with_default_argument,
     lambda_with_single_args,
+    pytest.param(
+        lambda_with_posonly_args,
+        marks=pytest.mark.skipif(not PY38, reason='posonly was added in 3.8'),
+    ),
     lambda_with_single_kwargs,
     lambda_with_single_kwonly,
 ])
@@ -43,6 +49,10 @@ def test_correct_arguments_count(
     lambda_with_single_argument,
     lambda_with_default_argument,
     lambda_with_single_args,
+    pytest.param(
+        lambda_with_posonly_args,
+        marks=pytest.mark.skipif(not PY38, reason='posonly was added in 3.8'),
+    ),
     lambda_with_single_kwargs,
     lambda_with_single_kwonly,
 ])
@@ -66,6 +76,10 @@ def test_correct_arguments_count_custom_option(
     lambda_with_single_argument,
     lambda_with_default_argument,
     lambda_with_single_args,
+    pytest.param(
+        lambda_with_posonly_args,
+        marks=pytest.mark.skipif(not PY38, reason='posonly was added in 3.8'),
+    ),
     lambda_with_single_kwargs,
     lambda_with_single_kwonly,
 ])