+++ wemake_python_styleguide/compat/functions.py
@@ -7,7 +7,7 @@ from wemake_python_styleguide.types import AnyAssign, AnyFunctionDefAndLambda
 
 
 def get_assign_targets(
-    node: Union[AnyAssign, ast.AugAssign],
+    node: Union[AnyAssign, ast.AugAssign],  # TODO: support NamedExpr
 ) -> List[ast.expr]:
     """Returns list of assign targets without knowing the type of assign."""
     if isinstance(node, (ast.AnnAssign, ast.AugAssign)):