+++ wemake_python_styleguide/visitors/ast/functions.py
@@ -2,7 +2,7 @@
 
 import ast
 from contextlib import suppress
-from typing import ClassVar, Dict, List, Union
+from typing import ClassVar, Dict, List, Mapping, Union
 
 from typing_extensions import final
 
@@ -12,7 +12,7 @@ from wemake_python_styleguide.constants import (
     FUNCTIONS_BLACKLIST,
     LITERALS_BLACKLIST,
 )
-from wemake_python_styleguide.logic import nodes, walk
+from wemake_python_styleguide.logic import nodes, source, walk
 from wemake_python_styleguide.logic.arguments import function_args
 from wemake_python_styleguide.logic.naming import access
 from wemake_python_styleguide.logic.tree import (
@@ -44,7 +44,11 @@ from wemake_python_styleguide.violations.refactoring import (
 )
 from wemake_python_styleguide.visitors import base, decorators
 
-LocalVariable = Union[ast.Name, ast.ExceptHandler]
+#: Things we treat as local variables.
+_LocalVariable = Union[ast.Name, ast.ExceptHandler]
+
+#: Function definitions with name and arity:
+_Defs = Mapping[str, int]
 
 
 @final
@@ -54,6 +58,21 @@ class WrongFunctionCallVisitor(base.BaseNodeVisitor):
 
     All these functions are defined in ``FUNCTIONS_BLACKLIST``.
     """
+
+    _functions: ClassVar[_Defs] = {
+        'getattr': 3,
+        'setattr': 3,
+    }
+
+    _postfixes: ClassVar[_Defs] = {
+        # dict methods:
+        '.get': 2,
+        '.pop': 2,
+        '.setdefault': 2,
+
+        # list methods:
+        '.insert': 2,
+    }
 
     def visit_Call(self, node: ast.Call) -> None:
         """
@@ -69,8 +88,12 @@ class WrongFunctionCallVisitor(base.BaseNodeVisitor):
         """
         self._check_wrong_function_called(node)
         self._check_boolean_arguments(node)
-        self._check_super_call(node)
         self._check_isinstance_call(node)
+
+        if functions.given_function_called(node, {'super'}):
+            self._check_super_context(node)
+            self._check_super_arguments(node)
+
         self.generic_visit(node)
 
     def _check_wrong_function_called(self, node: ast.Call) -> None:
@@ -83,20 +106,33 @@ class WrongFunctionCallVisitor(base.BaseNodeVisitor):
             )
 
     def _check_boolean_arguments(self, node: ast.Call) -> None:
-        # Calls with single boolean argument are allowed
         if len(node.args) == 1 and not node.keywords:
-            return
-        for arg in node.args:
-            if isinstance(arg, ast.NameConstant):
-                # We do not check for `None` values here:
-                if arg.value is True or arg.value is False:
-                    self.add_violation(
-                        BooleanPositionalArgumentViolation(
-                            arg, text=str(arg.value),
-                        ),
-                    )
+            return  # Calls with single boolean argument are allowed
 
-    def _ensure_super_context(self, node: ast.Call) -> None:
+        for arg in node.args:
+            if not isinstance(arg, ast.NameConstant):
+                continue
+
+            is_ignored = self._is_call_ignored(node)
+
+            # We do not check for `None` values here:
+            if not is_ignored and arg.value in {True, False}:
+                self.add_violation(
+                    BooleanPositionalArgumentViolation(
+                        arg, text=str(arg.value),
+                    ),
+                )
+
+    def _check_isinstance_call(self, node: ast.Call) -> None:
+        function_name = functions.given_function_called(node, {'isinstance'})
+        if not function_name or len(node.args) != 2:
+            return
+
+        if isinstance(node.args[1], ast.Tuple):
+            if len(node.args[1].elts) == 1:
+                self.add_violation(WrongIsinstanceWithTupleViolation(node))
+
+    def _check_super_context(self, node: ast.Call) -> None:
         parent_context = nodes.get_context(node)
         parent_node = nodes.get_parent(node)
 
@@ -112,30 +148,29 @@ class WrongFunctionCallVisitor(base.BaseNodeVisitor):
             grand_context = nodes.get_context(parent_context)
             if isinstance(grand_context, ast.ClassDef):
                 return
+
         self.add_violation(
             oop.WrongSuperCallViolation(node, text='not inside method'),
         )
 
-    def _ensure_super_arguments(self, node: ast.Call) -> None:
+    def _check_super_arguments(self, node: ast.Call) -> None:
         if node.args or node.keywords:
             self.add_violation(
                 oop.WrongSuperCallViolation(node, text='remove arguments'),
             )
 
-    def _check_super_call(self, node: ast.Call) -> None:
-        function_name = functions.given_function_called(node, {'super'})
-        if function_name:
-            self._ensure_super_context(node)
-            self._ensure_super_arguments(node)
+    def _is_call_ignored(self, node: ast.Call) -> bool:
+        call = source.node_to_string(node.func)
+        func_called = functions.given_function_called(
+            node, self._functions.keys(),
+        )
 
-    def _check_isinstance_call(self, node: ast.Call) -> None:
-        function_name = functions.given_function_called(node, {'isinstance'})
-        if not function_name or len(node.args) != 2:
-            return
-
-        if isinstance(node.args[1], ast.Tuple):
-            if len(node.args[1].elts) == 1:
-                self.add_violation(WrongIsinstanceWithTupleViolation(node))
+        return bool(
+            func_called and len(node.args) == self._functions[func_called],
+        ) or any(
+            call.endswith(post) and len(node.args) == self._postfixes[post]
+            for post in self._postfixes
+        )
 
 
 @final
@@ -216,7 +251,7 @@ class FunctionDefinitionVisitor(base.BaseNodeVisitor):
 
     def visit_any_function(self, node: AnyFunctionDef) -> None:
         """
-        Checks regular, lambda, and async functions.
+        Checks regular, ``lambda``, and ``async`` functions.
 
         Raises:
             UnusedVariableIsUsedViolation
@@ -230,7 +265,8 @@ class FunctionDefinitionVisitor(base.BaseNodeVisitor):
         self.generic_visit(node)
 
     def _check_unused_variables(self, node: AnyFunctionDef) -> None:
-        local_variables: Dict[str, List[LocalVariable]] = {}
+        local_variables: Dict[str, List[_LocalVariable]] = {}
+
         for body_item in node.body:
             for sub_node in ast.walk(body_item):
                 if isinstance(sub_node, (ast.Name, ast.ExceptHandler)):
@@ -238,10 +274,16 @@ class FunctionDefinitionVisitor(base.BaseNodeVisitor):
                     self._maybe_update_variable(
                         sub_node, var_name, local_variables,
                     )
+
         self._ensure_used_variables(local_variables)
 
     def _check_argument_default_values(self, node: AnyFunctionDef) -> None:
-        for arg in node.args.defaults:
+        all_defaults = filter(None, (
+            *node.args.defaults,
+            *node.args.kw_defaults,
+        ))
+
+        for arg in all_defaults:
             real_arg = operators.unwrap_unary_node(arg)
             parts = attributes.parts(real_arg) if isinstance(
                 real_arg, ast.Attribute,
@@ -264,9 +306,9 @@ class FunctionDefinitionVisitor(base.BaseNodeVisitor):
 
     def _maybe_update_variable(
         self,
-        sub_node: LocalVariable,
+        sub_node: _LocalVariable,
         var_name: str,
-        local_variables: Dict[str, List[LocalVariable]],
+        local_variables: Dict[str, List[_LocalVariable]],
     ) -> None:
         defs = local_variables.get(var_name)
         if defs is not None:
@@ -287,7 +329,7 @@ class FunctionDefinitionVisitor(base.BaseNodeVisitor):
 
     def _ensure_used_variables(
         self,
-        local_variables: Dict[str, List[LocalVariable]],
+        local_variables: Mapping[str, List[_LocalVariable]],
     ) -> None:
         for varname, usages in local_variables.items():
             for node in usages:
@@ -298,7 +340,7 @@ class FunctionDefinitionVisitor(base.BaseNodeVisitor):
                         ),
                     )
 
-    def _get_variable_name(self, node: LocalVariable) -> str:
+    def _get_variable_name(self, node: _LocalVariable) -> str:
         if isinstance(node, ast.Name):
             return node.id
         return getattr(node, 'name', '')